Rewrites Marvelite::API::Client for auto-discovery of available methods
Marvelite::API::Client, now only requires to know which routes its
injected Marvelite::API::Router has, to automatically define its
methods to connect to the Marvel Comics API.

This is a first draft in removing the dependency to manually declare
available methods for the client. Thanks to [Ismael
Marin](https://github.com/igmarin) for the idea and inspiration.