Updates Series method names in tests
Converts any reference to `#serie_FOO` to `#series_FOO` to match the method names that are in use. Also renames the series fixture files to match this convention. Finally, everywhere a single series id is referred to in the tests, I use the word 'series' instead of 'serie'. Note, all references to the `#serie` method and its fixture file remain as is.